Fixes for failure round 5

+++ emission/tests/netTests/TestVisualize.py
@@ -20,34 +20,36 @@ logging.basicConfig(level=logging.DEBUG)
 
 
 class TestVisualize(unittest.TestCase):
+    def setUp(self):
+        self.clearRelatedDb()
+        etc.setupRealExample(self,
+                             "emission/tests/data/real_examples/shankari_2015-aug-27")
+        eaicf.filter_accuracy(self.testUUID)
+        estfm.move_all_filters_to_data()
+        logging.info(
+            "After loading, timeseries db size = %s" % edb.get_timeseries_db().count())
+        self.day_start_ts = 1440658800
+        self.day_end_ts = 1440745200
+        self.day_start_dt = arrow.get(2015, 8, 27)
+        self.day_end_dt = arrow.get(2015, 8, 28)
 
-  def setUp(self):
-    self.clearRelatedDb()
-    etc.setupRealExample(self, "emission/tests/data/real_examples/shankari_2015-aug-27")
-    eaicf.filter_accuracy(self.testUUID)
-    estfm.move_all_filters_to_data()        
-    logging.info("After loading, timeseries db size = %s" % edb.get_timeseries_db().count())
-    self.day_start_ts = 1440658800
-    self.day_end_ts = 1440745200
-    self.day_start_dt = arrow.get(2015,8,27)
-    self.day_end_dt = arrow.get(2015,8,28)
-
-  def clearRelatedDb(self):
-    edb.get_timeseries_db().remove()
-    edb.get_place_db().remove()
-    edb.get_stop_db().remove()
-
-    edb.get_trip_new_db().remove()
-    edb.get_section_new_db().remove()
+    def clearRelatedDb(self):
+        edb.get_timeseries_db().remove({"user_id": self.testUUID})
+        edb.get_analysis_timeseries_db().remove({"user_id": self.testUUID})
 
 
-  def testCommutePopRoute(self):
-    points = visualize.range_mode_heatmap("MotionTypes.BICYCLING", self.day_start_dt.timestamp, self.day_end_dt.timestamp)
-    # I have to add test data with modes, I will do that tomorrow.
+    def testCommutePopRoute(self):
+        points = visualize.range_mode_heatmap("MotionTypes.BICYCLING",
+                                          self.day_start_dt.timestamp,
+                                          self.day_end_dt.timestamp)
+        # I have to add test data with modes, I will do that tomorrow.
 
-  def testBerkeleyPopRoute(self):
-    points = visualize.Berkeley_pop_route(self.day_start_dt.timestamp, self.day_end_dt.timestamp)
-    self.assertTrue(len(['latlng']) > 0)
+
+    def testBerkeleyPopRoute(self):
+        points = visualize.Berkeley_pop_route(self.day_start_dt.timestamp,
+                                          self.day_end_dt.timestamp)
+        self.assertTrue(len(['latlng']) > 0)
+
 
 if __name__ == '__main__':
     unittest.main()
